+++ src/main/java/cart/application/ProductService.java
@@ -1,7 +1,7 @@
 package cart.application;
 
 import cart.domain.Product;
-import cart.dao.ProductDao;
+import cart.repository.dao.ProductDao;
 import cart.dto.ProductRequest;
 import cart.dto.ProductResponse;
 import org.springframework.stereotype.Service;
