+++ src/main/java/cart/repository/dao/MemberDao.java
@@ -1,14 +1,12 @@
-package cart.dao;
+package cart.repository.dao;
 
 import cart.domain.Member;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.List;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
 
 @Repository
 public class MemberDao {
