+++ src/main/java/cart/ui/PageController.java
@@ -1,7 +1,7 @@
 package cart.ui;
 
 import cart.application.ProductService;
-import cart.dao.MemberDao;
+import cart.repository.dao.MemberDao;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
