+++ src/test/java/cart/integration/CartItemIntegrationTest.java
@@ -1,6 +1,6 @@
 package cart.integration;
 
-import cart.dao.MemberDao;
+import cart.repository.dao.MemberDao;
 import cart.domain.Member;
 import cart.dto.CartItemQuantityUpdateRequest;
 import cart.dto.CartItemRequest;
